+++ app/serializers/item_serializer.rb
@@ -1,12 +1,16 @@
 # frozen_string_literal: true
 
 class ItemSerializer < ApplicationSerializer
-  ATTRIBUTES = %i[id slug kind name data info homebrew has_description].freeze
+  ATTRIBUTES = %i[id slug kind name original_name data info homebrew has_description].freeze
 
   attributes(*ATTRIBUTES)
 
   def name
     translate(object.name)
+  end
+
+  def original_name
+    object.name['en']
   end
 
   def homebrew # rubocop: disable Naming/PredicateMethod