+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - rendering blocks for DH characters
 - creating homebrew items for DH characters
 - Select component with filtering opportunity
+- items search by original name
 
 ## [0.4.12] - 2026-03-02
 ### Modified